Harness AI Remediation to fix: Improper Limitation of a Pathname to a Restricted Directory ('Path Traversal')

+++ src/main/java/org/owasp/webgoat/lessons/pathtraversal/ProfileUploadBase.java
@@ -4,6 +4,8 @@
  */
 package org.owasp.webgoat.lessons.pathtraversal;
 
+import java.nio.file.Path;
+
 import static org.owasp.webgoat.container.assignments.AttackResultBuilder.failed;
 import static org.owasp.webgoat.container.assignments.AttackResultBuilder.informationMessage;
 import static org.owasp.webgoat.container.assignments.AttackResultBuilder.success;
@@ -67,7 +69,11 @@ public class ProfileUploadBase implements AssignmentEndpoint {
 
   @SneakyThrows
   protected File cleanupAndCreateDirectoryForUser(String username) {
-    var uploadDirectory = new File(this.webGoatHomeDirectory, "/PathTraversal/" + username);
+    var baseDir = new File(this.webGoatHomeDirectory, "/PathTraversal/");
+    var uploadDirectory = new File(baseDir.getCanonicalPath(), username);
+    if (!uploadDirectory.getCanonicalPath().startsWith(baseDir.getCanonicalPath() + File.separator)) {
+        throw new IOException("Invalid username directory path");
+    }
     if (uploadDirectory.exists()) {
       FileSystemUtils.deleteRecursively(uploadDirectory);
     }
